+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [0.2.0]
+### Changed
+- [List](docs/List.md) methods findAll, filter, and slice now return new Lists with the same sorter as the calling list 
+- [List](docs/List.md) uses a better default sorter with [default-compare](https://www.npmjs.com/package/default-compare)
+
+### Added
+- Method discardAll to [List](docs/List.md)
+- Method unique to [List](docs/List.md)
+
 ## [0.1.3] - 2019-02-26
 ### Added
 - Added the method addUnique to [List](docs/List.md)
@@ -20,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [List](docs/List.md)
 
+[0.2.0]: https://github.com/DarrenPaulWright/hord/compare/v0.1.3...v0.2.0
 [0.1.3]: https://github.com/DarrenPaulWright/hord/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/DarrenPaulWright/hord/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/DarrenPaulWright/hord/compare/v0.1.0...v0.1.1